+++ packages/account/PasskeyAccount.ts
@@ -1,6 +1,6 @@
 import * as ethers from "ethers"
 
-import number from "~packages/number"
+import number from "~packages/util/number"
 import type { BigNumberish, HexString } from "~typing"
 
 import type { Account, Call } from "./index"
