+++ packages/account/PasskeyOwnerP256.ts
@@ -1,7 +1,7 @@
 import { p256 } from "@noble/curves/p256"
 import * as ethers from "ethers"
 
-import byte from "~packages/byte"
+import byte from "~packages/util/byte"
 import type { BytesLike } from "~typing"
 
 import type { PasskeyOwner } from "./PasskeyOwner"
