+++ packages/util/number/index.ts
